calyx_frontend::ast
Struct NamespaceDef

pub struct NamespaceDef {
    pub imports: Vec<PosString>,
    pub components: Vec<ComponentDef>,
    pub externs: Vec<(Option<PosString>, Vec<Primitive>)>,
    pub metadata: Option<String>,
    pub source_info_table: Option<SourceInfoTable>,
}

Corresponds to an individual Calyx file.

Fields
imports: Vec<PosString>

Path to extern files.

components: Vec<ComponentDef>

List of component definitions.

externs: Vec<(Option<PosString>, Vec<Primitive>)>

Extern statements and any primitive declarations in them.

metadata: Option<String>

Optional opaque metadata.

source_info_table: Option<SourceInfoTable>

Optional Source Info table

Implementations

impl NamespaceDef

pub fn construct(file: &Option<PathBuf>) -> CalyxResult<Self>

Construct a namespace from a file or the input stream. If no file is provided, the input stream must be a TTY.

pub fn construct_from_str(inp: &str) -> CalyxResult<Self>

Construct a namespace from a definition using a string.