calyx_frontend

Struct Attributes

pub struct Attributes { /* private fields */ }

Attributes associated with a specific IR structure.

Implementations

impl Attributes

pub fn insert<A>(&mut self, key: A, val: u64)
where A: Into<Attribute>,

Add a new attribute

pub fn insert_set<S>(&mut self, key: S, val: u32)
where S: Into<SetAttribute>,

pub fn get<A>(&self, key: A) -> Option<u64>
where A: Into<Attribute>,

Get the value associated with an attribute key

pub fn get_set<S>(&self, key: S) -> Option<&VecSet<u32>>
where S: Into<SetAttribute>,

pub fn has<A>(&self, key: A) -> bool
where A: Into<Attribute>,

Check if an attribute key has been set

pub fn is_empty(&self) -> bool

Returns true if there are no attributes

pub fn remove<A>(&mut self, key: A)
where A: Into<Attribute>,

Remove attribute with the name key

pub fn copy_from<A>(&mut self, other: Self, keys: Vec<A>)
where A: Into<Attribute> + Clone,

self copys (i.e., assigns the same values) the attributes in other. However, we only copy attributes in keys (i.e.. we donโt copy all attributes in other, only the ones that we specify). If a key is not present in other, then we ignore that key. Example: suppose self: A->10, B->5 other: A->15, C->5 keys: A, D Then self gets: A->15 B->5. (D is ignored since itโs not present in other and C is ignored since itโs not keys.)

pub fn copy_from_set<A>(&mut self, other: &Self, keys: Vec<A>)
where A: Into<SetAttribute> + Clone,

Copies the values of the given set attributes,keys, from other into self. Note that this does not overwrite set values in self that are already present.

pub fn add_span(self, span: GPosIdx) -> Self

Set the span information

pub fn to_string_with<F, S>( &self, sep: &'static str, fmt: F, set_fmt: S, ) -> String
where F: Fn(String, u64) -> String, S: Fn(String, &[u32]) -> String,